ensure test coverage is >80%, ensure test parallelization is utilized, and tests are in correct files
